+++ front/package.json
@@ -6,6 +6,7 @@
     "@emotion/react": "^11.7.1",
     "@emotion/styled": "^11.6.0",
     "@mui/icons-material": "^5.2.5",
+    "@mui/lab": "^5.0.0-alpha.65",
     "@mui/material": "^5.2.5",
     "@mui/styled-engine": "^5.2.5",
     "@testing-library/jest-dom": "^5.15.1",
@@ -14,6 +15,8 @@
     "@tinymce/tinymce-react": "^3.13.0",
     "axios": "^0.24.0",
     "buffer": "^6.0.3",
+    "formik": "^2.2.9",
+    "formik-mui": "^4.0.0-alpha.3",
     "prop-types": "^15.8.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
@@ -23,7 +26,8 @@
     "react-router-dom": "^6.2.1",
     "react-scripts": "^5.0.0",
     "uuid": "^8.3.2",
-    "web-vitals": "^1.1.2"
+    "web-vitals": "^1.1.2",
+    "yup": "^0.32.11"
   },
   "devDependencies": {
     "eslint": "^8.5.0",
